pyproject.toml
[tool.poetry]
name = "py-pdf-recibos"
version = "0.1.0"
description = "Gerador de recibos em PDF usando Python e FPDF"
authors = ["Alisson R. Perez <756802+alissonperez@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fpdf2 = "^2.8.1"
fire = "^0.7.0"

[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
